+++ packages/ruby/ext/kreuzberg_rb/extconf.rb
@@ -24,4 +24,5 @@ default_profile = ENV.fetch('CARGO_PROFILE', 'release')
 
 create_rust_makefile('kreuzberg_rb') do |config|
   config.profile = default_profile.to_sym
+  config.ext_dir = File.expand_path('native', __dir__)
 end